aws-backend/model-server/requirements.txt
fastapi
uvicorn[standard]
pillow
torch
torchvision
